flowernet-outliner/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.6.1
google-genai==1.62.0
